Fixed tag updating

+++ commands/updatetags.js
@@ -18,6 +18,7 @@ module.exports = {
         let gid = message.guild.id;
 
         let updateCount = 0;
+        let doneCount = 0;
         for(let m in message.guild.members.array()) {
             let newMember = message.guild.members.array()[m];
             let newTag = '';
@@ -25,8 +26,8 @@ module.exports = {
 
             newMember.roles.array().forEach(function (role) {
                 if ((!guildSettings[gid] || !guildSettings[gid].checkhoist || role.hoist) && role.position > pos) {
-                    pos = role.position;
-                    newTag = getTagForRole(role.id, newMember.guild.roles);
+                    newTag = getTagForRole(role.id, newMember.guild.roles) || newTag;
+                    pos = getTagForRole(role.id, newMember.guild.roles) ? role.position : pos;
                 }
             });
 
@@ -34,36 +35,64 @@ module.exports = {
             if (!nickname)
                 nickname = newMember.user.username;
             let memberTag = nickname.split(' ')[0].replace('[', '').replace(']', '');
+            if(nickname.indexOf('[') !== 0) {
+                memberTag = "";
+            }
             if (getRoleForTag(memberTag, newMember.guild.roles)) {
-                nickname = nickname.substring(memberTag.length+3, nickname.length);
+                nickname = nickname.substring(memberTag === "" ? 0 : (memberTag.length+3), nickname.length);
             }
 
-            if (memberTag === newTag) {
-                return;
+            if ("[" + memberTag.toUpperCase() + "]" === newTag || (memberTag === "" && newTag === "")) {
+                continue;
             }
 
             updateCount++;
 
             newMember.edit({
                 nick: newTag + ' ' + nickname
+            }).then(function() {
+                doneCount++;
+                if(doneCount === updateCount) {
+                    let addText = "That's a lot. I'm glad I cleaned that mess up!";
+                    if(updateCount <= 50) {
+                        addText = " Quite a few actually!";
+                    }
+                    if(updateCount <= 10) {
+                        addText = "Not too many.";
+                    }
+                    if(updateCount === 0) {
+                        addText = "Seems like everything was already tidied up!";
+                    }
+                    message.channel.send("Done updating role tags! I had to update " + updateCount + " role tags! " + addText);
+                    message.delete();
+                }
             }).catch(function (e) {
                 logger.warn(e.message);
+                updateCount--;
+                if(doneCount === updateCount) {
+                    let addText = "That's a lot. I'm glad I cleaned that mess up!";
+                    if(updateCount <= 50) {
+                        addText = " Quite a few actually!";
+                    }
+                    if(updateCount <= 10) {
+                        addText = "Not too many.";
+                    }
+                    if(updateCount === 0) {
+                        addText = "Seems like everything was already tidied up!";
+                    }
+                    message.channel.send("Done updating role tags! I had to update " + updateCount + " role tags! " + addText);
+                    message.delete();
+                }
             });
         }
-
-        let addText = "That's a lot. I'm glad I cleaned that mess up!";
-        if(updateCount <= 50) {
-            addText = " Quite a few actually!";
+        if(updateCount === doneCount) {
+            let addText = "";
+            if (updateCount === 0) {
+                addText = "Seems like everything was already tidied up!";
+            }
+            message.channel.send("Done updating role tags! I had to update " + updateCount + " role tags! " + addText);
+            message.delete();
         }
-        if(updateCount <= 10) {
-            addText = "Not too many.";
-        }
-        if(updateCount === 0) {
-            addText = "Seems like everything was already tidied up!";
-        }
-
-        message.channel.send("Done updating role tags! I had to update " + updateCount + " role tags! " + addText);
-        message.delete();
     },
     help: ""
 };
@@ -86,9 +115,9 @@ function getRoleForTag(text, roles) {
 function getTagForRole(role, roles) {
     for(let entry of roles) {
         let r = entry[1];
-        let tagCloserPos = r.name.substring(3,5).indexOf(']');
-        if(r.id === role) {
-            let roleTag = r.name.substring(1, 3+tagCloserPos).toLowerCase();
+        let tagCloserPos = r.name.indexOf(']');
+        if(r.id === role && tagCloserPos && r.name.indexOf('[') === 0) {
+            let roleTag = r.name.substring(1, tagCloserPos).toLowerCase();
             return "[" + roleTag.toUpperCase() + "]";
         }
     }
